tictacthor revamp
