+++ Accounts/tictacthor/clients/src/main/java/net/corda/samples/tictacthor/webserver/Starter.java
@@ -1,0 +1,23 @@
+package net.corda.samples.tictacthor.webserver;
+
+import org.springframework.boot.Banner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import static org.springframework.boot.WebApplicationType.SERVLET;
+
+/**
+ * Our Spring Boot application.
+ */
+@SpringBootApplication
+public class Starter {
+    /**
+     * Starts our Spring Boot application.
+     */
+    public static void main(String[] args) {
+        SpringApplication app = new SpringApplication(Starter.class);
+        app.setBannerMode(Banner.Mode.OFF);
+        app.setWebApplicationType(SERVLET);
+        app.run(args);
+    }
+}
