+++ Accounts/tictacthor/contracts/build.gradle
@@ -32,6 +32,6 @@ dependencies {
     cordaCompile "$corda_core_release_group:corda-core:$corda_core_release_version"
     cordaRuntime "$corda_release_group:corda:$corda_release_version"
     testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-    // Token Account dependencies.
+    // Account dependencies.
     cordaCompile "$accounts_release_group:accounts-contracts:$accounts_release_version"
 }
