--- Accounts/tictacthor/contracts/src/test/java/com/tictacthor/contracts/ContractTests.java
@@ -1,13 +1,0 @@
-package com.tictacthor.contracts;
-
-import net.corda.testing.node.MockServices;
-import org.junit.Test;
-
-public class ContractTests {
-    private final MockServices ledgerServices = new MockServices();
-
-    @Test
-    public void dummyTest() {
-
-    }
-}
