+++ Accounts/tictacthor/gradle.properties
@@ -1,3 +1,3 @@
-name=Test
+name=Tictacthor Cordapp
 group=com.tictacthor
-version=0.1
+version=1.0
