+++ Accounts/tictacthor/runServers.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+./gradlew bootRunDevRel &
+./gradlew bootRunSOE
