+++ Accounts/tictacthor/workflows/build.gradle
@@ -57,6 +57,7 @@ dependencies {
 
     // CorDapp dependencies.
     cordapp project(":contracts")
+
     //Account dependencies
     cordaCompile "$confidential_id_release_group:ci-workflows:$confidential_id_release_version"
     cordaCompile "$accounts_release_group:accounts-workflows:$accounts_release_version"
