--- Accounts/tictacthor/workflows/src/test/java/com/tictacthor/ContractTests.java
@@ -1,13 +1,0 @@
-package com.tictacthor;
-
-import net.corda.testing.node.MockServices;
-import org.junit.Test;
-
-public class ContractTests {
-    private final MockServices ledgerServices = new MockServices();
-
-    @Test
-    public void dummyTest() {
-
-    }
-}
