--- Accounts/tictacthor/workflows/src/test/java/com/tictacthor/FlowTests.java
@@ -1,37 +1,0 @@
-package com.tictacthor;
-
-import com.google.common.collect.ImmutableList;
-import net.corda.testing.node.MockNetwork;
-import net.corda.testing.node.MockNetworkParameters;
-import net.corda.testing.node.StartedMockNode;
-import net.corda.testing.node.TestCordapp;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-public class FlowTests {
-    private final MockNetwork network = new MockNetwork(new MockNetworkParameters(ImmutableList.of(
-        TestCordapp.findCordapp("com.tictacthor.contracts"),
-        TestCordapp.findCordapp("com.tictacthor.flows")
-    )));
-    private final StartedMockNode a = network.createNode();
-    private final StartedMockNode b = network.createNode();
-
-    public FlowTests() {
-    }
-
-    @Before
-    public void setup() {
-        network.runNetwork();
-    }
-
-    @After
-    public void tearDown() {
-        network.stopNodes();
-    }
-
-    @Test
-    public void dummyTest() {
-
-    }
-}
